Move pfs related functions to pfs.c/pfs.h
