+++ main.c
@@ -46,6 +46,7 @@
 #include "coredump.h"
 #include "usb.h"
 #include "qr.h"
+#include "pfs.h"
 
 int _newlib_heap_size_user = 128 * 1024 * 1024;
